+++ dtslint/ts3.5/ReaderEither.ts
@@ -87,7 +87,7 @@ pipe(
 // do notation
 //
 
-// $ExpectType ReaderEither<{ readonly a: number; } & { readonly b: string; }, string | number, { a: number; b: string; c: boolean; }>
+// $ExpectType ReaderEither<{ readonly a: number; } & { readonly b: string; }, string | number, { readonly a: number; readonly b: string; readonly c: boolean; }>
 pipe(
   _.right<{ readonly a: number }, string, number>(1),
   _.bindTo('a'),
@@ -99,7 +99,7 @@ pipe(
 // pipeable sequence S
 //
 
-// $ExpectType ReaderEither<{ readonly a: number; } & { readonly b: string; }, string | number, { a: number; b: string; c: boolean; }>
+// $ExpectType ReaderEither<{ readonly a: number; } & { readonly b: string; }, string | number, { readonly a: number; readonly b: string; readonly c: boolean; }>
 pipe(
   _.right<{ readonly a: number }, string, number>(1),
   _.bindTo('a'),
@@ -111,7 +111,7 @@ pipe(
 // Do
 //
 
-// $ExpectType ReaderEither<unknown, string, { a: number; b: string; }>
+// $ExpectType ReaderEither<unknown, string, { readonly a: number; readonly b: string; }>
 pipe(
   _.Do,
   _.bind('a', () => _.of<unknown, string, number>(1)),