+++ dtslint/ts3.5/ReaderTaskEither.ts
@@ -109,7 +109,7 @@ pipe(
 // do notation
 //
 
-// $ExpectType ReaderTaskEither<{ readonly a: number; } & { readonly b: string; }, string | number, { a: number; b: string; c: boolean; }>
+// $ExpectType ReaderTaskEither<{ readonly a: number; } & { readonly b: string; }, string | number, { readonly a: number; readonly b: string; readonly c: boolean; }>
 pipe(
   _.right<{ readonly a: number }, string, number>(1),
   _.bindTo('a'),
@@ -121,7 +121,7 @@ pipe(
 // pipeable sequence S
 //
 
-// $ExpectType ReaderTaskEither<{ readonly a: number; } & { readonly b: string; }, string | number, { a: number; b: string; c: boolean; }>
+// $ExpectType ReaderTaskEither<{ readonly a: number; } & { readonly b: string; }, string | number, { readonly a: number; readonly b: string; readonly c: boolean; }>
 pipe(
   _.right<{ readonly a: number }, string, number>(1),
   _.bindTo('a'),
@@ -133,7 +133,7 @@ pipe(
 // Do
 //
 
-// $ExpectType ReaderTaskEither<unknown, string, { a: number; b: string; }>
+// $ExpectType ReaderTaskEither<unknown, string, { readonly a: number; readonly b: string; }>
 pipe(
   _.Do,
   _.bind('a', () => _.of<unknown, string, number>(1)),