+++ src/TaskEither.ts
@@ -58,6 +58,7 @@ import { MonadIO2 } from './MonadIO'
 import { MonadTask2, MonadTask2C } from './MonadTask'
 import { MonadThrow2, MonadThrow2C } from './MonadThrow'
 import { Monoid } from './Monoid'
+import { Option } from './Option'
 import { Pointed2 } from './Pointed'
 import { Predicate } from './Predicate'
 import { Refinement } from './Refinement'
@@ -875,7 +876,7 @@ export const FromEither: FromEither2<URI> = {
  * @category constructors
  * @since 2.0.0
  */
-export const fromOption =
+export const fromOption: <E>(onNone: Lazy<E>) => <A>(ma: Option<A>) => TaskEither<E, A> =
   /*#__PURE__*/
   fromOption_(FromEither)
 
@@ -906,7 +907,7 @@ export const fromTaskOption: <E>(onNone: Lazy<E>) => <A>(e: TaskOption<A>) => Ta
  * @category combinators
  * @since 2.4.0
  */
-export const chainEitherK =
+export const chainEitherK: <E, A, B>(f: (a: A) => E.Either<E, B>) => (ma: TaskEither<E, A>) => TaskEither<E, B> =
   /*#__PURE__*/
   chainEitherK_(FromEither, Chain)
 
@@ -924,7 +925,10 @@ export const chainEitherKW: <E2, A, B>(
  * @category constructors
  * @since 2.0.0
  */
-export const fromPredicate =
+export const fromPredicate: {
+  <E, A, B extends A>(refinement: Refinement<A, B>, onFalse: (a: A) => E): (a: A) => TaskEither<E, B>
+  <E, A>(predicate: Predicate<A>, onFalse: (a: A) => E): (a: A) => TaskEither<E, A>
+} =
   /*#__PURE__*/
   fromPredicate_(FromEither)
 
@@ -932,7 +936,10 @@ export const fromPredicate =
  * @category combinators
  * @since 2.0.0
  */
-export const filterOrElse =
+export const filterOrElse: {
+  <E, A, B extends A>(refinement: Refinement<A, B>, onFalse: (a: A) => E): (ma: TaskEither<E, A>) => TaskEither<E, B>
+  <E, A>(predicate: Predicate<A>, onFalse: (a: A) => E): (ma: TaskEither<E, A>) => TaskEither<E, A>
+} =
   /*#__PURE__*/
   filterOrElse_(FromEither, Chain)
 
@@ -953,7 +960,9 @@ export const filterOrElseW: {
  * @category combinators
  * @since 2.4.0
  */
-export const fromEitherK =
+export const fromEitherK: <E, A extends ReadonlyArray<unknown>, B>(
+  f: (...a: A) => E.Either<E, B>
+) => (...a: A) => TaskEither<E, B> =
   /*#__PURE__*/
   fromEitherK_(FromEither)
 
@@ -1143,7 +1152,7 @@ export const bindW: <N extends string, A, E2, B>(
   f: (a: A) => TaskEither<E2, B>
 ) => <E1>(
   fa: TaskEither<E1, A>
-) => TaskEither<E1 | E2, { [K in keyof A | N]: K extends keyof A ? A[K] : B }> = bind as any
+) => TaskEither<E1 | E2, { readonly [K in keyof A | N]: K extends keyof A ? A[K] : B }> = bind as any
 
 // -------------------------------------------------------------------------------------
 // pipeable sequence S
@@ -1164,7 +1173,7 @@ export const apSW: <A, N extends string, E2, B>(
   fb: TaskEither<E2, B>
 ) => <E1>(
   fa: TaskEither<E1, A>
-) => TaskEither<E1 | E2, { [K in keyof A | N]: K extends keyof A ? A[K] : B }> = apS as any
+) => TaskEither<E1 | E2, { readonly [K in keyof A | N]: K extends keyof A ? A[K] : B }> = apS as any
 
 // -------------------------------------------------------------------------------------
 // array utils